вндерил страницы в общий лэйаут
