--- packages/client/src/components/forum-thread/forum-thread.module.pcss
@@ -1,37 +1,0 @@
-.container {
-  padding-top: var(--gap);
-}
-
-.panel {
-  padding: 24px;
-  max-width: 930px;
-  margin: 0 auto;
-  border-radius: 12px;
-  background-color: var(--color-light-primary);
-  height: calc(100vh - 2 * var(--gap));
-}
-
-.link {
-  color: var(--color-accent-primary);
-  padding-bottom: 24px;
-  display: inline-flex;
-  align-items: center;
-}
-
-.text {
-  line-height: 30px;
-  margin-left: 5px;
-}
-
-.thread {
-  padding: 0 73px;
-  width: 100%;
-}
-
-.title {
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 24px;
-  color: var(--color-dark-primary);
-  padding-bottom: 15px;
-}
