+++ packages/client/src/pages/forum-thread/forum-thread.module.pcss
@@ -1,0 +1,35 @@
+.container {
+  padding: var(--gap);
+  display: flex;
+  flex-direction: column;
+  gap: var(--gap);
+}
+
+.link {
+  color: var(--color-accent-primary);
+  display: inline-flex;
+  align-items: center;
+}
+
+.text {
+  margin-left: 7px;
+}
+
+.wrapper {
+  padding-left: 72px;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  min-height: 0; /* Необходимо для правильной работы скролла */
+}
+
+.posts {
+  padding-right: 60px;
+}
+
+.title {
+  font-size: 20px;
+  font-weight: 700;
+  line-height: 24px;
+  color: var(--color-dark-primary);
+}
