+++ packages/client/src/pages/forum-thread/index.ts
